refactor: split the monolithic index.ts into modules by concern
The single 1106-line src/index.ts is broken into small modules grouped by
concern, each helper folder documented with a descriptive .md:

  config.ts / state.ts / peers.ts   core: settings, SQLite state, P2P signing
  immich/    local Immich API, refs, contributors, proxy writes
  p2p/       cross-server wire protocol + join
  sync/      reconciliation + comment loops
  media/     hotlink byte path, LRU cache, app-facing interceptor
  web/       HTML pages, HTTP router, passthrough, banner

index.ts is now a thin composition root; web/server.ts is a thin dispatch
table (the byte-interceptor and Immich passthrough moved to media/ and web/).

Behaviour is unchanged — code was moved verbatim (dependency DAG is cycle-free;
p2p/sync reference each other only at call time). Proven identical by the full
e2e suite (66/66) against the rebuilt image. Dockerfile now copies the whole
src tree; release-please tracks the version in config.ts. ARCHITECTURE.md gains
a code-layout map + doc conventions, and its sync description is corrected to
nudge-driven-with-a-timed-backstop (no websockets).
