+++ release-please-config.json
@@ -5,7 +5,9 @@
   "packages": {
     ".": {
       "changelog-path": "CHANGELOG.md",
-      "extra-files": ["src/index.ts"]
+      "extra-files": [
+        "src/config.ts"
+      ]
     }
   }
 }
